+++ src/com/DTO/FundPurchaseDTO.java
@@ -19,7 +19,7 @@ public class FundPurchaseDTO extends BaseVO {
     private Long fundId;
     private Long customerId;
     private String purchaseDate;
-    private int num;
+    private double num;
     private FundDTO fundDTO;
 
     public FundDTO getFundDTO() {
@@ -30,11 +30,11 @@ public class FundPurchaseDTO extends BaseVO {
         this.fundDTO = fundDTO;
     }
 
-    public int getNum() {
+    public double getNum() {
         return num;
     }
 
-    public void setNum(int num) {
+    public void setNum(double num) {
         this.num = num;
     }
